moves animal and breed selections to own  classes
